Created a styled component: RelatedCarousel to separate styled and function components for RelatedProducts
